core/ai/openai_provider.py
import os
from openai import OpenAI
from dotenv import load_dotenv



class OpenAIProvider:
    def __init__(self):
        api_key = os.getenv("OPENAI_API_KEY")

        if not api_key:
            raise ValueError("OPENAI_API_KEY not found in environment variables.")

        self.client = OpenAI(api_key=api_key)

    def generate(self, prompt: str) -> str:
        response = self.client.chat.completions.create(
            model="gpt-4o-mini",
            messages=[
                {
                    "role": "system",
                    "content": (
                        "You are a software generation engine. "
                        "Always return valid JSON with this structure: "
                        "{ \"files\": [ { \"path\": \"string\", \"content\": \"string\" } ] }"
                    ),
                },
                {
                    "role": "user",
                    "content": prompt,
                },
            ],
            temperature=0.3,
        )

        return response.choices[0].message.content