requirements.txt
requests
json-repair
python-dotenv
flask
flask-cors
fastapi
uvicorn
sqlalchemy
pytest
python-telegram-bot==21.9
